initialized expo app
