+++ front-end/.gitignore
@@ -9,3 +9,6 @@ npm-debug.*
 *.orig.*
 web-build/
 web-report/
+
+# macOS
+.DS_Store
