+++ front-end/app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "PooNU",
-    "slug": "poo-nu-app",
+    "name": "poo-nu",
+    "slug": "poo-nu",
     "privacy": "public",
     "sdkVersion": "35.0.0",
     "platforms": [
@@ -11,9 +11,9 @@
     ],
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./assets/images/icon.png",
+    "icon": "./assets/icon.png",
     "splash": {
-      "image": "./assets/images/splash.png",
+      "image": "./assets/splash.png",
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
     },
