+++ front-end/babel.config.js
