--- front-end/poo-nu-app/.expo-shared/assets.json
@@ -1,6 +1,0 @@
-{
-  "0cae4d70c6df3e5e96ee8b5c442b59d55c8ab8deb466992ab9abc523822f2a1b": true,
-  "e997a5256149a4b76e6bfd6cbf519c5e5a0f1d278a3d8fa1253022b03c90473b": true,
-  "af683c96e0ffd2cf81287651c9433fa44debc1220ca7cb431fe482747f34a505": true,
-  "e7fc0741cc6562975a990e3d9ef820571588dab20aba97032df9f00caa9cd57a": true
-}
