--- front-end/poo-nu-app/.watchmanconfig
@@ -1,1 +1,0 @@
-{}
