--- front-end/poo-nu-app/package.json
@@ -1,35 +1,0 @@
-{
-  "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject",
-    "test": "jest --watchAll"
-  },
-  "jest": {
-    "preset": "jest-expo"
-  },
-  "dependencies": {
-    "@expo/samples": "~3.0.3",
-    "@expo/vector-icons": "^10.0.3",
-    "@react-navigation/web": "^1.0.0-alpha.9",
-    "expo": "^35.0.0",
-    "expo-asset": "^7.0.0",
-    "expo-constants": "^7.0.0",
-    "expo-font": "^7.0.0",
-    "expo-web-browser": "^7.0.0",
-    "react": "16.8.3",
-    "react-dom": "16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
-    "react-native-gesture-handler": "~1.3.0",
-    "react-native-web": "^0.11.7",
-    "react-navigation": "^3.12.0"
-  },
-  "devDependencies": {
-    "babel-preset-expo": "^7.0.0",
-    "jest-expo": "^35.0.0"
-  },
-  "private": true
-}
